+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.0 (master)
 
 * Instruction graphs can be enabled with `pyjion.enable_graphs()` and then exported by `pyjion.get_graph(f)`
+* Pyjion will raise an ImportError if .NET is missing instead of a system exit
 
 ## 1.0.0 (beta2)